pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omnispindle"
version = "0.1.0"
description = "A distributed task management engine for orchestrating agent tasks"
readme = "README.md"
authors = [
    {name = "Madness Interactive Team"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "redis>=4.3.4",
    "pydantic>=1.9.0",
    "python-dateutil>=2.8.2",
    "aiohttp>=3.8.1",
    "click>=8.1.3",
    "prometheus-client>=0.14.1",
    "python-dotenv>=0.20.0",
    "protobuf>=4.21.0",
    "websockets>=10.3",
    "SQLAlchemy>=1.4.39"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.3",
    "black>=22.3.0",
    "isort>=5.10.1",
    "mypy>=0.961",
    "flake8>=4.0.1",
    "pre-commit>=2.19.0",
    "faker>=13.12.0"
]
postgres = [
    "psycopg2>=2.9.3",
    "asyncpg>=0.25.0"
]
docs = [
    "sphinx>=5.0.1",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-autodoc-typehints>=1.18.3"
]

[project.scripts]
omnispindle = "omnispindle.cli:main"

[tool.setuptools]
packages = ["omnispindle"]

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--strict-markers -v" 